+++ src/com/company/repository/DatabaseAdapter.java
@@ -20,7 +20,7 @@ public class DatabaseAdapter implements Database{
 
 
     @Override
-    public void saveGame(String username, HashMap<AtomType, Integer> atomMap, HashMap<PowerUpType, Integer> powerUpMap, HashMap<PowerUpType, Integer> userPowerUpMap, HashMap<MoleculeType, Integer> moleculeMap, Map<ReactionBlockerType, Integer> reactionBlockerAmount, int score, boolean isLinear, int time) {
+    public void saveGame(String username, HashMap<AtomType, Integer> atomMap, HashMap<PowerUpType, Integer> powerUpMap, HashMap<PowerUpType, Integer> userPowerUpMap, HashMap<MoleculeType, Integer> moleculeMap, Map<ReactionBlockerType, Integer> reactionBlockerAmount, double score, boolean isLinear, int time) {
         database.saveGame(username, atomMap, powerUpMap, userPowerUpMap, moleculeMap, reactionBlockerAmount, score,isLinear, time);
     }
